pyproject.toml
[project]
name = "learn-litellm-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "instructor>=1.14.5",
    "litellm>=1.81.7",
    "python-dotenv>=1.2.1",
]
